+++ randovania/gui/lib/qt_network_client.py
@@ -12,7 +12,7 @@ import randovania
 from randovania.gui.dialog.text_prompt_dialog import TextPromptDialog
 from randovania.gui.lib import async_dialog, wait_dialog
 from randovania.network_client.network_client import ConnectionState, NetworkClient, UnableToConnect
-from randovania.network_common import error, server_signals
+from randovania.network_common import error
 from randovania.network_common.async_race_room import AsyncRaceRoomEntry
 from randovania.network_common.multiplayer_session import (
     MultiplayerSessionActions,
@@ -22,6 +22,7 @@ from randovania.network_common.multiplayer_session import (
     MultiplayerWorldPickups,
     WorldUserInventory,
 )
+from randovania.network_common.signals import server_signals
 from randovania.network_common.user import CurrentUser
 
 if TYPE_CHECKING:
